+++ src/main/java/com/springapp/exception/EntityNotFoundException.java
@@ -7,17 +7,17 @@ import com.springapp.jpa.model.EntityId;
  */
 public class EntityNotFoundException extends RuntimeException implements EntityId {
 
-    private final Long entityId;
+    private final Integer entityId;
     private final String entityName;
 
-    public EntityNotFoundException(final Long entityId,
+    public EntityNotFoundException(final Integer entityId,
                                    final String entityName) {
         this.entityId = entityId;
         this.entityName = entityName;
     }
 
     @Override
-    public Long getId() {
+    public Integer getId() {
         return entityId;
     }
