+++ src/main/java/com/springapp/jpa/model/Exercise.java
@@ -29,7 +29,7 @@ public class Exercise implements EntityId {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "exercise_id")
-    private Long id;
+    private Integer id;
 
     @NotNull
     @Size(min = 5, max = 20)
@@ -45,11 +45,11 @@ public class Exercise implements EntityId {
     @OneToMany(fetch = FetchType.LAZY, mappedBy = "pk.user", cascade = CascadeType.REMOVE)
     private Set<UserExercise> userExercise = new HashSet<>();
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(final Long id) {
+    public void setId(final Integer id) {
         this.id = id;
     }
